# backend/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Audit Wolf API
  description: |
    Comprehensive smart contract auditing platform API that combines AI-powered analysis 
    with traditional static analysis tools to provide detailed security assessments of Solidity contracts.
  version: 1.0.0
  contact:
    name: Audit Wolf Support
    email: support@auditwolf.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.auditwolf.com/api
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /analysis/start:
    post:
      tags:
        - Analysis
      summary: Start contract analysis
      description: Queue a new analysis job for a contract
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartAnalysisRequest"
      responses:
        "201":
          description: Analysis started successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisStartResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /analysis/{auditId}/progress:
    get:
      tags:
        - Analysis
      summary: Get analysis progress
      description: Get real-time progress of an ongoing analysis
      parameters:
        - name: auditId
          in: path
          required: true
          description: Audit UUID
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Progress retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgressResponse"
        "404":
          description: Audit not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /analysis/{auditId}/results:
    get:
      tags:
        - Analysis
      summary: Get analysis results
      description: Retrieve completed analysis results
      parameters:
        - name: auditId
          in: path
          required: true
          description: Audit UUID
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Results retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResultsResponse"
        "404":
          description: Audit not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /analysis/validate:
    post:
      tags:
        - Analysis
      summary: Validate contract for analysis
      description: Quick validation of contract source code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateContractRequest"
      responses:
        "200":
          description: Validation completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Error message"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string

    # Health Schemas
...
    ValidateContractRequest:
      type: object
      required:
        - sourceCode
      properties:
        sourceCode:
          type: string
          example: "pragma solidity ^0.8.0;\n\ncontract MyToken {\n    // Contract code here\n}"
    ValidationResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            isValid:
              type: boolean
            errors:
              type: array
              items:
                type: string
            metrics:
              type: object
              properties:
                complexity:
                  type: integer
                lines_of_code:
                  type: integer
                functions_count:
                  type: integer

    # Analysis Schemas
    StartAnalysisRequest:
      type: object
      required:
        - contractId
        - analysisType
      properties:
        contractId:
          type: string
          format: uuid
        analysisType:
          type: string
          enum: [static, ai, full]
        priority:
          type: integer
          enum: [1, 5, 10, 15]
          description: "1=low, 5=normal, 10=high, 15=critical"
        options:
          type: object
          properties:
            includeGasOptimization:
              type: boolean
            aiModels:
              type: array
              items:
                type: string
    AnalysisStartResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            auditId:
              type: string
              format: uuid
            jobId:
              type: string
            message:
              type: string
    ProgressResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            status:
              type: string
              enum: [pending, analyzing, completed, failed]
            progress:
              type: integer
              minimum: 0
              maximum: 100
            currentStep:
              type: string
            estimatedTimeRemaining:
              type: integer
              description: "Seconds remaining"
    AnalysisResultsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            auditId:
              type: string
              format: uuid
            status:
              type: string
            vulnerabilities:
              type: array
              items:
                $ref: "#/components/schemas/Vulnerability"
            gasOptimizations:
              type: array
              items:
                $ref: "#/components/schemas/GasOptimization"
            summary:
              type: object
              properties:
                total_vulnerabilities:
                  type: integer
                critical_count:
                  type: integer
                high_count:
                  type: integer
                medium_count:
                  type: integer
                low_count:
                  type: integer
...
    Vulnerability:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum:
            [
              reentrancy,
              overflow,
              access_control,
              gas_optimization,
              best_practice,
            ]
        severity:
          type: string
          enum: [critical, high, medium, low, informational]
        title:
          type: string
        description:
          type: string
        location:
          type: object
          properties:
            line:
              type: integer
            column:
              type: integer
            function:
              type: string
        recommendation:
          type: string
        confidence:
          type: number
          minimum: 0
          maximum: 1
        source:
          type: string
          enum: [static, ai, combined]
    GasOptimization:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        location:
          type: object
          properties:
            line:
              type: integer
            function:
              type: string
        estimated_savings:
          type: integer
        recommendation:
          type: string

    # Report Schemas
...